requirements.txt
chainlit
langchain
langchain-community
chromadb
firecrawl
python-dotenv
requests
langchain
chromadb
python-dotenv